WRITE_EXTERNAL_STORAGE required for downloads
