+++ app/src/main/AndroidManifest.xml
@@ -3,6 +3,7 @@
           xmlns:android="http://schemas.android.com/apk/res/android">
     <uses-permission android:name="android.permission.INTERNET"/>
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
     <!-- TODO: test with RTL layouts -->
     <application
         android:name=".Derpibooru"
